disable lookups when game's rendering mode breaks the lookup UI
Lookup Anything uses nested sprite batches for its rendering logic, and this always works except in very rare cases when the game uses SpriteSortMode.Immediate. This is only known to happen in the Stardew Valley Fair, and the only thing you can lookup in that event is yourself. Given that, the alternative approaches would greatly complicate the rendering logic for little benefit.