ChestsAnywhere: add option to stack items
This patch extends the ChestsAnywhere mod to allow the player to push a
button and have all the items in her inventory that already have stacks
in her chests be "stacked" with them.

I've noticed that quite a lot of my time playing the game is spent on
putting each item in the respective chest after looting, which is
irritating to the point that spoils the enjoyment of the game (to me, at
least.)

With this option, it only takes a single click (or keypress) to combine
items found in the inventory with existing stacks that have been placed
in the chest before-hand.

Future Improvements
-------------------

I'd like to be able to assign "specialties" to each chest, like
Vegetables or Fish, and have the stacking functionality utilize this
piece of data such that it can stack items even if they are not present
in the chest already. This would be useful for new seasons where you're
foraging and fishing new items, in which case it will reduce the manual
labor.

Problems
--------

- I couldn't find a proper place for the button so right now it's to the
  side of the existing Organize button
- For some reason, the hover text is not being honored in the chest
  menu; it would be nice to have a tooltip for the button since it may
  not be very intuitive. (Granted, this seems to be the case for the
  existing Sort/Organize button, so maybe we can fix both...)