Respect quality of items when stacking
This commit fixes an issue where the item stacker would confuse items of
the same type but of different "star ratings". It also moves the
stacking logic to a separate class that is now interacted with by both
the overlay (for the button) and the main mod file (for the key binding)
so that stacking can be done even when the menu is not open.