Fix compatibility between tractor mod & harvest with scythe mod.
The problem is that Harvest with scythe slightly modifies the meaning of `Crop.harvestMethod.Value`. This results in the tractor mod to call `HoeDirt.performUseAction` on a crop that subsequently is considered a `Crop.sickleHarvest` crop causing it to trigger a harvest with scythe animation. By calling `dirt.crop.harvest` directly this problem is circumvented.