Lookup Anything: Show construction recipes that need an item as an ingredient.
Add Building Input recipe type.
Update building recipes data parser to be the Building Input type.
Gather construction recipes in GameHelper.
The item created by a recipe can now be null - check for null in ItemSubject.
Add SpecialOutput to RecipeModel to allow for output that is not based on an Item.