[Content Patcher] Changing "English name" to "internal name" in CP docs in "tokens.md"
Hi!

(First time creating a pull request for a repo that is not mine, so hopefully I did it correctly!
)
I was checking the docs for Content Patcher for my mod, and I noticed that in the tokens.md, the tokens "Hearts", "Relationship", "Roommate", and "Spouse" were referring to the NPC's English name. However, the code is (or, at least, looks like it does; haven't checked that, as my C# knowledge is very poor) checking for the NPC's internal name, no matter what the display name (https://stardewvalleywiki.com/Modding:NPC_data#Basic_info) is. Therefore, I changed that in the docs, so instead of "(using their English name regardless of translations)", it will be "(using their internal name regardless of translations)".

Feel free to adapt, change, not accept, or do whatever you want with my pull request!

StrojvedouciDenis